requirements.txt
Keras==2.2.4
matplotlib==3.0.2
moviepy==1.0.1
numpy==1.16.1
pandas==0.25.1
Pillow==6.2.0
requests==2.22.0
tensorflow==1.15.0
